requirements.txt
# Core dependencies
fastmcp==2.13.0.2
pinecone==7.3.0
ollama==0.6.0
python-dotenv==1.2.1
pytest==8.4.2
